+++ packages/frontend/src/features/omnibar/utils/getSearchItemLabel.ts
@@ -14,6 +14,8 @@ export const getSearchItemLabel = (itemType: SearchItemType) => {
             return 'Tables';
         case SearchItemType.PAGE:
             return 'Pages';
+        case SearchItemType.SQL_CHART:
+            return 'SQL Charts';
         default:
             return assertUnreachable(
                 itemType,
@@ -33,6 +35,7 @@ export const getSearchItemErrorLabel = (itemType: SearchItemType) => {
         case SearchItemType.SPACE:
         case SearchItemType.TABLE:
         case SearchItemType.PAGE:
+
         default:
             return new Error(`Unknown error item type: ${itemType}`);
     }